Added things

+++ app/src/main/java/com/wiseowl/notifier/domain/model/Location.kt
@@ -1,6 +1,6 @@
 package com.wiseowl.notifier.domain.model
 
 data class Location(
-    val longitude: Float,
-    val latitude: Float
+    val longitude: Double,
+    val latitude: Double
 )
